// dataclarity-api/src/modules/indicators/enums/indicator-target-status.enum.ts
export enum IndicatorTargetStatus {
  // HIGHER_IS_BETTER / LOWER_IS_BETTER
  TARGET_ACHIEVED = 'TARGET_ACHIEVED',
  ON_TRACK = 'ON_TRACK',
  AT_RISK = 'AT_RISK',
  OFF_TRACK = 'OFF_TRACK',
  // RANGE_IS_BETTER
  WITHIN_RANGE = 'WITHIN_RANGE',
  BELOW_RANGE = 'BELOW_RANGE',
  ABOVE_RANGE = 'ABOVE_RANGE',
  // Sem dados
  NO_GOAL = 'NO_GOAL',
  NO_DATA = 'NO_DATA',
}